+++ packages/api/lambdas/sf-starter.js
@@ -78,11 +78,6 @@ async function incrementAndDispatch(queueUrl, queueMessage) {
     return await dispatch(queueUrl, queueMessage);
   } catch (error) {
     await decrementQueueSemaphore(queueUrl);
-
-    if (error.code === 'ExecutionAlreadyExists' || error.message.includes('ExecutionAlreadyExists')) {
-      logger.info('Execution already exists. Proceeding to delete message from queue...');
-      await sqs.deleteSQSMessage(queueUrl, queueMessage.ReceiptHandle);
-    }
     throw error;
   }
 }